playing with casl, strict null, logging prisma queries

+++ src/modules/user/user.controller.ts
@@ -1,4 +1,4 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete, UseInterceptors, ParseIntPipe, UseGuards, } from '@nestjs/common';
+import { Controller, Get, Post, Body, Patch, Param, Delete, UseInterceptors, ParseIntPipe, UseGuards, NotFoundException, } from '@nestjs/common';
 import { UserService } from './user.service';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
@@ -33,6 +33,9 @@ export class UserController {
     @Body() updateUserDto: UpdateUserDto,
     @Ability() ability: AppAbility) {
     const subjectUser = await this.userService.findById(id);
+    if (!subjectUser) {
+      throw new NotFoundException();
+    }
     ForbiddenError.from(ability).throwUnlessCan(Action.Update, subject('User', subjectUser));
 
     return this.userService.update(id, updateUserDto);
@@ -42,6 +45,9 @@ export class UserController {
   @UseGuards(CaslGuard)
   async remove(@Param('id', ParseIntPipe) id: number, @Ability() ability: AppAbility) {
     const subjectUser = await this.userService.findById(id);
+    if (!subjectUser) {
+      throw new NotFoundException();
+    }
     ForbiddenError.from(ability).throwUnlessCan(Action.Delete, subject('User', subjectUser));
 
     return this.userService.remove(id);
